Revert "chore: Split up api.ts"
