--- src/api/index.ts
@@ -1,38 +1,0 @@
-export * from "./hooks";
-
-export {
-  getProblemsXlsx,
-  deleteMedia,
-  moveMedia,
-  getElevation,
-  getGradeDistribution,
-  getPermissions,
-  getSvgEdit,
-  getUserSearch,
-  getUsersTicks,
-  postComment,
-  postPermissions,
-  postProblem,
-  postProblemMedia,
-  postProblemSvg,
-  postSector,
-  postTicks,
-  postUserRegion,
-  putMediaInfo,
-  putMediaJpegRotate,
-} from "./operations";
-
-export {
-  getLocales,
-  getBaseUrl,
-  useAccessToken,
-  getImageUrl,
-  getBuldreinfoMediaUrlSupported,
-  getBuldreinfoMediaUrl,
-  getAreaPdfUrl,
-  getSectorPdfUrl,
-  getProblemPdfUrl,
-  numberWithCommas,
-  convertFromDateToString,
-  convertFromStringToDate,
-} from "./utils";
