--- src/api/types.ts
@@ -1,5 +1,0 @@
-import { ConsistencyAction } from "../components/DataReloader";
-
-export type FetchOptions = Partial<Parameters<typeof fetch>[1]> & {
-  consistencyAction?: ConsistencyAction;
-};
